+++ app/src/main/java/com/example/googleauthlogin/model/CartItem.kt
@@ -4,6 +4,6 @@ class CartItem (
     val cartItemId: String = "",
     val cartItemName: String = "",
     val cartItemImg: String = "",
-    val cartItemCost: Long = 0,
+    val cartItemCost: Double = 0.0,
     val quantity: Int = 0
 )
